Revert "Add threading to webcam"
This reverts commit e069ad51b3f8bd0c54aaea934c122db72be34ded.

This was done because the threading does not noticeably improve the fps of the
video feed. This implies that the slow fps is due to the processing of the
images, and not due to slow image reading. Therefore fps can only be improved by
making the processing more efficient or by threading it.